app/requirements.txt
streamlit==1.32.0
pandas==2.0.3
gspread==5.12.0
google-auth==2.28.1
google-auth-oauthlib==1.0.0
gspread-dataframe==3.3.1
plotly==5.18.0